修改全量迁移承接数据总行数的字段类型int为long

+++ src/main/java/org/opengauss/portalcontroller/entity/Total.java
@@ -25,7 +25,7 @@ public class Total {
     /**
      * 所有表的总记录数，预估值
      */
-    private int record;
+    private long record;
     /**
      * 所有表的数据总量，预估值
      */
